backend/requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
sentry-sdk==1.38.0
pydantic==2.5.0
pydantic-settings==2.1.0
psycopg2-binary==2.9.9
sqlmodel==0.0.21
email-validator==2.2.0
boto3==1.34.0
python-docx==1.1.0
mammoth==1.6.0
html2docx==1.6.0
